feat: Add redirect option to signInWithPassword
Add a new property 'redirect' to the signInWithPassword function in the auth.js file. This property allows the user to specify whether they want to be redirected after signing in. By default, the value is set to 'true', but now it can be set to 'false' to prevent redirection.

This commit addresses the need to provide more flexibility in the sign-in process and improves the user experience by allowing them to control the redirection behavior.